requirements.txt
streamlit==1.29.0
openai==1.12.0
nltk==3.8.1
google==3.0.0
googlesearch-python==1.2.3
goose3==3.1.19
yake==0.4.8
python-dotenv